Improved stats function
command minifab stats only works when network is
up running. When the network is not running, it
will display an error message which does not work
well in some automated environment. This PR will
fix that issue and display network health percentage
based on how many nodes are running healthy.

Signed-off-by: Tong Li <litong01@us.ibm.com>