Changed connection file to use embeded certs
connection files have been using referenced certs in the past
for golang connection profile, it is not very convinient since
there will be many files to be moved with connection files.
This PR fixed that issue so that all certs will come with
the connection file. Also added example using wallet.

Signed-off-by: Tong Li <litong01@us.ibm.com>